Tests/LibWeb/Text/expected/selection-drag-past-overflowing-text.txt

kept label selected below overflowing text: true
